[update] 補正種類の整理

+++ src/Computer.h
@@ -54,7 +54,7 @@ private:
     void copyAgent();
     void resetCopyTile();
     int conflictMove(int x, int y, unsigned int agentNum, int angle);
-    int distance(MoveData currentMoveData);
+    double distance(MoveData currentMoveData);
     int decodeCorrection(int num);
     void splitTurn();
     void partSelect();
